+++ src/DynamicDefinition.jl
@@ -3,13 +3,11 @@ defines generic Dynamic type
 """
 
 module DynamicDefinition
-
-using ValidatedNumerics
-using TaylorSeries:Taylor1
 export Dynamic, MarkovDynamic, preim, nbranches, plottable, is_full_branch, domain, derivative, distorsion, endpoints, branch, expansivity, max_distorsion, orientation
 
 abstract type Dynamic end
 abstract type MarkovDynamic <: Dynamic end
+using IntervalArithmetic, IntervalOptimisation
 
 domain(S::Dynamic) = @error "Not implemented"
 nbranches(S::Dynamic) = @error "Not implemented"
@@ -22,6 +20,7 @@ preim(S::Dynamic, k, y, ϵ)
 Computes the preim of y in branch k of a dynamic, with accuracy ϵ
 """
 function preim end
+
 is_full_branch(S::Dynamic) = @error "Not implemented"
 
 """
@@ -30,13 +29,15 @@ Endpoints of the branches, in increasing order (returned as a vector of interval
 endpoints(S::Dynamic) = @error "Not implemented"
 
 # Derivative and distorsion of a generic function (*not* a dynamic). Here for convenience,
-
 # the isempty check is required because otherwise derivative(x -> 4*x, ∅) == 4.
+
+#
+import TaylorSeries
 """
 Nth derivative of a function (or a dynamic)
 """
 derivative(f, x) = derivative(1, f, x)
-derivative(n, f, x) = isempty(x) ?  ∅ : f(Taylor1([x, 1], n))[n] * factorial(n)
+derivative(n, f, x) = isempty(x) ?  ∅ : f(TaylorSeries.Taylor1([x, 1], n))[n] * factorial(n)
 
 """
 Distorsion of a function (or a dynamic), i.e., |f′′ / f′^2|
@@ -45,7 +46,7 @@ function distorsion(f, x)
 	if isempty(x)
 		return ∅
 	end
-	series = f(Taylor1([x, 1], 2))
+	series = f(TaylorSeries.Taylor1([x, 1], 2))
 	f′ = series[1]
 	f′′ = 2*series[2]
 	return abs(f′′ / f′^2)
